refactor interval_join_table
