+++ bodo/ir/join.py
@@ -39,7 +39,6 @@ from bodo.libs.array import (
     cpp_table_to_py_data,
     delete_table,
     hash_join_table,
-    interval_join_table,
     nested_loop_join_table,
     py_data_to_cpp_table,
 )
@@ -1280,6 +1279,87 @@ def _gen_cross_join_repeat(
     nodes = f_block.body[:-3]
     nodes[-1].target = join_node.out_data_vars[0]
     return nodes
+
+
+@intrinsic
+def interval_join_table(
+    typingctx,
+    left_table_t,
+    right_table_t,
+    left_parallel_t,
+    right_parallel_t,
+    is_left_t,
+    is_right_t,
+    is_left_point_t,
+    is_strict_contains_t,
+    is_strict_left_t,
+    point_col_id_t,
+    interval_start_col_id_t,
+    interval_end_col_id_t,
+    key_in_output_t,
+    need_typechange_t,
+    _bodo_rebalance_output_if_skewed,
+    num_rows_ptr_t,
+):
+    """
+    Call cpp function for optimized interval join of two tables.
+    Point in interval and interval overlap joins are supported.
+    """
+    from bodo.libs.array import table_type
+
+    assert left_table_t == table_type, "interval_join_table: cpp table type expected"
+    assert right_table_t == table_type, "interval_join_table: cpp table type expected"
+
+    def codegen(context, builder, sig, args):
+        fnty = lir.FunctionType(
+            lir.IntType(8).as_pointer(),
+            [
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(1),
+                lir.IntType(64),
+                lir.IntType(64),
+                lir.IntType(64),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(1),
+                lir.IntType(8).as_pointer(),
+            ],
+        )
+        fn_tp = cgutils.get_or_insert_function(
+            builder.module, fnty, name="interval_join_table"
+        )
+        ret = builder.call(fn_tp, args)
+        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
+        return ret
+
+    return (
+        table_type(
+            left_table_t,
+            right_table_t,
+            types.boolean,
+            types.boolean,
+            types.boolean,
+            types.boolean,
+            types.boolean,
+            types.boolean,
+            types.boolean,
+            types.uint64,
+            types.uint64,
+            types.uint64,
+            types.voidptr,
+            types.voidptr,
+            types.boolean,
+            types.voidptr,
+        ),
+        codegen,
+    )
 
 
 def join_distributed_run(
